Use column names to find indices in mapped and unmapped.
We're adding new columns, and expecting the columns do be located at specific indices mean this will only work for reports from a particular version.

Also, The column that was meant to be Destination Series in th GCP Discounts table, was by mistake picking up the column Source Shape instead. This is now fixed.